+++ malachite-base/src/u64_polynomial/arithmetic/mod.rs
@@ -12,6 +12,11 @@ pub mod height;
 /// An implementation of [`ModIsReduced`](crate::num::arithmetic::traits::ModIsReduced), which
 /// checks whether every coefficient of a polynomial is less than a given modulus.
 pub mod mod_is_reduced;
+/// Implementations of [`Mod`](crate::num::arithmetic::traits::Mod),
+/// [`ModAssign`](crate::num::arithmetic::traits::ModAssign), [`Rem`](core::ops::Rem) and
+/// [`RemAssign`](core::ops::RemAssign), which reduce every coefficient of a polynomial modulo a
+/// number.
+pub mod mod_op;
 /// Implementations of [`ModPowerOf2`](crate::num::arithmetic::traits::ModPowerOf2) and
 /// [`ModPowerOf2Assign`](crate::num::arithmetic::traits::ModPowerOf2Assign), which reduce every
 /// coefficient of a polynomial modulo a power of 2.